requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
requests==2.30.0
six==1.16.0
sniffio==1.3.0
tabulate==0.9.0
tzdata==2023.3
urllib3==2.0.2
